Fix bug making some files not output correctly
Sometimes out was not being set to the correct value and causing some
templates which have multiple variants to overwrite the first output
path in the yaml file.

Removing the option to specify your own output path and creating a let
statement solved this problem.

In future the functionality may be restored but as this function was
added to implement the base16 builder spec which does not allow builders to
specify their own path this hack was chosen.

A possible way to return this is to check if we have a output path
supplied and then set template-output-path to that or construct the
path from the template yaml otherwise and then set that to
template-output-path.